Tidy

+++ src/main/java/me/eccentric_nz/TARDIS/destroyers/TARDISPruner.java
@@ -142,7 +142,7 @@ public class TARDISPruner {
     }
 
     private long getTime(int day) {
-        long period = (long) (day * 86400000L);
+        long period = day * 86400000L;
         long now = System.currentTimeMillis();
         long prune = now - period;
         return prune;
